No longer fetching lyrics snippets from MusixMatch
This behavior is now optional; must give --deep. It really doesn't help
that much, since the snippets don't always exist anyway, nor are they
always accurate. So since we already have to rely on various filters to
work around this, there's not much since wasting a bunch of API calls